gcp/requirements.txt
tensorflow==2.10.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.1.3
google-cloud-storage==2.14.0
Pillow==9.4.0
fastapi==0.111.0
uvicorn==0.23.2
starlette==0.37.2
functions-framework==3.9.1

